+++ src/test/java/PageModel/PlayListPage.java
@@ -7,6 +7,9 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class PlayListPage extends Basepage{
     public PlayListPage(WebDriver givenDriver) {
         super(givenDriver);
@@ -51,6 +54,31 @@ public class PlayListPage extends Basepage{
     public void clickAddToPlaylistBtn() {
         WebElement addToBtn = driver.findElement(By.cssSelector("[data-test='add-to-btn']"));
         addToBtn.click();
+    }
+    public boolean isPlistNameInHeader(String playlist) {
+        WebElement playlistHeader = driver.findElement(By.cssSelector("#playlistWrapper h1"));
+        return wait.until(ExpectedConditions
+                .textToBePresentInElement(playlistHeader, playlist));
+
+    }
+
+    public void clickToDeletePlist() {
+        WebElement deletePlaylistBtn = wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector(".btn-delete-playlist")));
+        deletePlaylistBtn.click();
+
+    }
+
+    public List<String> getPlaylistNames() {
+        List<WebElement> playlists = driver.findElements(By.cssSelector("#playlists a"));
+        // get playlist names from playlist elements
+        List<String> playlistNames = new ArrayList<>();
+
+        for (int i = 0; i < playlists.size(); i++) {
+            String playlistName = playlists.get(i).getText();
+            playlistNames.add(playlistName);
+        }
+        System.out.println(playlistNames);
+        return playlistNames;
     }
 }
